Create README.md
I'm currently a Helpdesk Analyst. I've been in this field for over 3 years now. At this time, there's not an opportunity for growth within the company that I work for. I've been researching careers that would be a good fit for beginners in the tech world. There's a lot of buzz about cloud computing so I'm taking the introduction to cloud computing on AWS right now. After gaining the knowledge of how the cloud actually works, I plan to take the AWS Certified Cloud Practitioner Complete course.